refactor: WhiteboardUseCase 리팩터링
- 기존 WhiteboardObjectUseCase에서 하는 기능을 모두 담당합니다.
- 추가로 화이트보드 연결 해제 기능을 제공합니다.
